Refactored the dashboard activity components and styles.
