+++ src/pages/ActivityContainer/DashBoardActivity.scss
@@ -1,0 +1,99 @@
+@import '../../baseStyle.scss';
+
+.dashboard-content-wrapper {
+  .card-activity-container {
+    @include flexLayoutWithGap(row, center, flex-start, 1rem);
+    width: 100%;
+    height: 385px;
+
+    .cardWrapper {
+      width: 53%;
+      height: 91%;
+
+      .card-head__header {
+        width: 100%;
+        height: 40px;
+        border-bottom: 1px solid #d3d3d3;
+      }
+
+      .chart-container {
+        width: 100%;
+        height: calc(100% - 40px);
+      }
+
+    }
+
+    .card-pie-chart-container {
+      @include flexLayoutWithGap(column-reverse, center, flex-start, 1rem);
+      width: 40%;
+      height: 100%;
+
+      .cardWrapper {
+        width: 93%;
+        height: 39.5%;
+
+        .pie-chart-container {
+          @include flexLayoutWithGap(row, center, center, 1rem);
+          width: 100%;
+          height: 100%;
+          font-size: 12px;
+        }
+
+        .timeline-chart-container {
+          @include flexLayoutWithGap(row, center, flex-start, 1rem);
+          width: 100%;
+          height: 100%;
+          font-size: 10px;
+
+          .time-line {
+            height: 100px;
+            padding: 5px 0;
+
+            .red-line {
+              border: 1px solid #f01d0d;
+              width: 0;
+              height: 32.5px;
+              margin-bottom: 1px;
+            }
+
+            .green-line {
+              border: 1px solid #199608;
+              width: 0;
+              height: 32.5px;
+              margin-bottom: 1px;
+            }
+
+            .blue-line {
+              border: 1px solid #02149b;
+              width: 0;
+              height: 32.5px;
+              margin-bottom: 1px;
+            }
+          }
+
+          .time-line__info {
+            height: 100px;
+
+            .red-line__info {
+              @include flexLayoutWithGap(column, flex-start, center, .3rem);
+              height: 34.5px;
+              margin-bottom: 1px;
+            }
+
+            .green-line__info {
+              @include flexLayoutWithGap(column, flex-start, center, .3rem);
+              height: 34.5px;
+              margin-bottom: 1px;
+            }
+
+            .blue-line__info {
+              @include flexLayoutWithGap(column, flex-start, center, .3rem);
+              height: 34.5px;
+              margin-bottom: 1px;
+            }
+          }
+        }
+      }
+    }
+  }
+}
