+++ src/pages/Dashboard/DashboardContent.jsx
@@ -6,12 +6,11 @@ import './Dashboard.scss';
 import { FiPaperclip, FiClipboard } from 'react-icons/fi';
 import { BiLockAlt, BiBlock } from 'react-icons/bi';
 import { RiBarChart2Fill } from 'react-icons/ri';
-import CardWithChart from './CardWithChart';
-import DashboardActivity from './DashboardActivity';
 
 import Table from '../../components/Table/Table';
 import DashboardToDoList from './DashboardToDoList';
 import SalesGraph from './SalesGraph';
+import DashBoardActivity from '../ActivityContainer/DashBoardActivity';
 
 const DashboardContent = () => {
   const renderIcon = value => {
@@ -108,10 +107,7 @@ const DashboardContent = () => {
           </Card>
         </div>
       </div>
-      <div className='card-activity-container'>
-        <DashboardActivity />
-        <CardWithChart />
-      </div>
+      <DashBoardActivity />
     </div>
   );
 };
